requirements.txt
Flask==2.2.2
Keras_Preprocessing==1.1.2
matplotlib==3.5.2
numpy==1.23.1
opencv_python==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
pydicom==2.3.0
scikit_learn==1.1.2
seaborn==0.11.2
tensorflow==2.9.1
tqdm==4.64.0
